Extract out unittest for colorable
Updated it to pass the tests and changed something so added in a quick fix
